Improve inviting players
Invites can now be given players who do not yet have account in clan
database, but exist in dgamelaunch database.  This necessitated two
extensions: Searching for eligible invitees now searches both dgamelaunch
and clan databases and the results are merged together.  Additionally,
when player is given invite and they don't have an account in clandb yet,
they are automatically instantiated as if the logged in themselves.